+++ packages/hippo_server/linux_host_inspector/lib/linux_host_inspector.dart
@@ -5,3 +5,4 @@ export 'src/file_system.dart';
 export 'src/inspector.dart';
 export 'src/models.dart';
 export 'src/native_system.dart';
+export 'src/system_metrics.dart';